Fix deque_any_cast_element
