+++ tests/regression/deque_any_cast_element/deque_any_cast_element.t.cpp
@@ -1,9 +1,9 @@
 #include <cstdio>
 #include <cassert>
-#include "DequeAnyCastElement.h"
+#include "deque_any_cast_element.h"
 
 int main() {
-    uint64_t r = DequeAnyCastElement::test_result;
+    uint64_t r = test_result;
     assert(r == 3);
     printf("test passed\n");
     return 0;
