+++ tests/regression/dune
@@ -5937,3 +5937,14 @@
   (alias runtest)
   (deps deque_any_cast.t.exe)
   (action (run ./deque_any_cast.t.exe))))
+
+(subdir deque_any_cast_element
+ (rule
+  (targets deque_any_cast_element.t.exe)
+  (deps DequeAnyCastElement.vo deque_any_cast_element.t.cpp (source_tree .))
+  (action
+   (run %{project_root}/scripts/compile-std.sh %{project_root} deque_any_cast_element.t.exe deque_any_cast_element.cpp deque_any_cast_element.t.cpp)))
+ (rule
+  (alias runtest)
+  (deps deque_any_cast_element.t.exe)
+  (action (run ./deque_any_cast_element.t.exe))))
