--- tests/wip/deque_any_cast_element/DequeAnyCastElement/Specif.cpp
@@ -1,3 +1,0 @@
-#include "Specif.h"
-
-namespace Specif {} // namespace Specif
