--- tests/wip/deque_any_cast_element/DequeAnyCastElement/Specif.h
@@ -1,22 +1,0 @@
-#ifndef INCLUDED_SPECIF
-#define INCLUDED_SPECIF
-
-#include <utility>
-
-namespace Specif {
-
-template <typename A, typename P> struct SigT {
-  // DATA
-  A x;
-  P a1;
-
-  // ACCESSORS
-  SigT<A, P> clone() const { return {x, a1}; }
-
-  // CREATORS
-  static SigT<A, P> existt(A x, P a1) { return {std::move(x), std::move(a1)}; }
-};
-
-} // namespace Specif
-
-#endif // INCLUDED_SPECIF
