+++ tests/wip/dune
@@ -112,14 +112,3 @@
   (alias runtest)
   (deps type_args_extraction.t.exe)
   (action (run ./type_args_extraction.t.exe))))
-
-(subdir deque_any_cast_element
- (rule
-  (targets deque_any_cast_element.t.exe)
-  (deps DequeAnyCastElement.vo deque_any_cast_element.t.cpp (source_tree .))
-  (action
-   (run %{project_root}/scripts/compile-std.sh %{project_root} deque_any_cast_element.t.exe DequeAnyCastElement.cpp deque_any_cast_element.t.cpp)))
- (rule
-  (alias runtest)
-  (deps deque_any_cast_element.t.exe)
-  (action (run ./deque_any_cast_element.t.exe))))
